state estimation applied
The number of states is now estimated and all immediately needed data tables are allocated for that number.
The later needed auxiliary variables are allocated for the exact number of states (which is known when this allocations happens).
The MAX_STATE macro is no more used in the practical code. It is only used with compile time parameter to determine the STATE datatype.
Allphabet optimization is now always done (regardless of the goto function type).
LINKSQ_ARRAY optimization macro is disabled.
The pointer table used for goto function is now allocated at once.